Added mergedConfig
