+++ processor/transformer/transformer.go
@@ -46,6 +46,7 @@ type MetadataT struct {
 	TrackingPlanId      string                 `json:"trackingPlanId"`
 	TrackingPlanVersion int                    `json:"trackingPlanVersion"`
 	SourceTpConfig      map[string]interface{} `json:"sourceTpConfig"`
+	MergedTpConfig      map[string]interface{} `json:"mergedTpConfig"`
 	DestinationID       string                 `json:"destinationId"`
 	JobRunID            string                 `json:"jobRunId"`
 	JobID               int64                  `json:"jobId"`
